style: Applied styling to TextButton component
It has two types of styling one from within component and other being passed as props.
If styles passed to comonents are empty then the default empty object is used in place of styles passed as props.

First/ styles created within the component are always applied unless the same styling properties are overddien by the same styling properties used/specifies in the styles passed as props.