feat: Created History component to get all data for calender from the api
Saved data received from the API to the redux store.
Added messsage for the day if there is no entry data for the day.
Connected the History component with the redux store using connec() of 'react-redux'.
Created mapStateToProps() to get data saved in the redux store to use it on the component