feat: Display actual data instead of stringigy version
Added DateHeader component to show date as header.
Styled the component.
Applied touchableopacity button to the day where data was logged.
For currentday/today if no data is logged then an message to log data is displayed.
For all other days if there is no data for that day then message saying that you didnt log any data is showed.