feat: Implemented the SetLocation() function which is used to set the watch the user location
It also sets/updates the local state with provided  response from Location API.

Added componentDidMount() lifecycle hook which is used to ask location permissions.
If permissions granted then set them in local state as granted and call the setLocation() to start watching user's location. Otherwise set it to whatever the permission status is.

If an error occurs during geeting the permission then display error and set the status of location services permission to 'undetermined'