feat: Applied the local notification to application
In App.js applied the setlocalNotification() within componentDidMount() lifecycle hook so that we can ask user to get the permisiions to set notofications on device when app loads and sets the notification as well.

In AddEntry component when ever new entry is added ( using submit() function),  we clear the previous notification ( using clearLocalNotification function ) which was set in App.js componentDidMount() because we dont need to display this notification now because user has already set the data for this day.

Notification for current day will only be cleared when user add new entry for current day because clearing the otification function is called in side submit() used to add new entry for the day.

Once the notification is cleared, we also set new notification for tmoorow using setLocalNotification() inside submit() and after clearLocalNotofication()